requirements.txt
requests>=2.31
praw>=7.7
python-dotenv>=1.0
PySide6>=6.5
Pillow>=10.0
